requirements.txt
amqp==5.0.2
appnope==0.1.2
asgiref==3.3.1
autopep8==1.5.4
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Django==3.1.5
django-3-jet==1.0.8
django-autoslug==1.9.8
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-rosetta==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
polib==1.1.0
prompt-toolkit==3.0.10
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.6.0
Pygments==2.7.4
PyJWT==2.0.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stories==4.0.0
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.2
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
